# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HỆ THỐNG CỬA HÀNG ĐIỆN THOẠI PHÚC MOBILE",
        "description": "Hệ thống API quản lý cửa hàng kinh doanh điện thoại di động và phụ kiện công nghệ PHÚC MOBILE. Hỗ trợ đầy đủ phân quyền người dùng (RBAC) và chuẩn hóa dữ liệu quan hệ 3NF.\n *     \n *     ### Danh sách Vai trò (Roles):\n *     - **superadmin**: Quyền cao nhất, quản lý toàn bộ hệ thống bao gồm cả nhân viên và khách hàng.\n *     - **admin**: Quản lý nghiệp vụ chính (Điện thoại, Hãng sản xuất, Danh mục, Mã giảm giá, Đơn hàng, Đánh giá).\n *     - **staff**: Nhân viên vận hành cửa hàng, xử lý trạng thái đơn hàng và phê duyệt đánh giá sản phẩm.\n *     - **user**: Khách hàng (Xem sản phẩm, thêm giỏ hàng, đặt mua điện thoại và gửi đánh giá).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server (Tự động nhận diện Host)"
        }
    ],
    "paths": {
...
        "/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Thanh toán đơn hàng (Checkout)",
                "operationId": "6ed91804beb738578b8d4c98f822dbbf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "receiver_name",
                                    "receiver_phone",
                                    "province_id",
                                    "district_id",
                                    "ward_id",
                                    "address_detail",
                                    "payment_method"
                                ],
                                "properties": {
                                    "receiver_name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "receiver_phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "example": 202
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "example": 1442
                                    },
                                    "ward_id": {
                                        "type": "integer",
                                        "example": 20109
                                    },
                                    "address_detail": {
                                        "type": "string",
                                        "example": "Số 123, Đường ABC"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cod",
                                            "bank_transfer",
                                            "momo",
                                            "vnpay"
                                        ],
                                        "example": "cod"
                                    },
                                    "coupon_code": {
                                        "type": "string",
                                        "example": "SUMMER2024",
                                        "nullable": true
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Giao gio hanh chinh",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đặt hàng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đặt hàng thành công!"
                                        },
                                        "order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validation hoặc mã giảm giá không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Giỏ hàng của bạn đang trống."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Danh sách đơn hàng của người dùng hiện tại",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Chi tiết đơn hàng",
                "operationId": "f088d34a2737782215f435f854e31de5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Hủy đơn hàng",
                "description": "Chỉ được hủy khi trạng thái là pending",
                "operationId": "f01ac571088d34f82562b3d48911f6f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hủy thành công"
                    },
                    "422": {
                        "description": "Không thể hủy đơn hàng ở trạng thái hiện tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/orders": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "[Admin] Danh sách toàn bộ đơn hàng",
                "operationId": "36165b6ece5d97e1fad9531988955d13",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "[Admin] Chi tiết đơn hàng",
                "operationId": "4152a373b1695abb81538ce82956bd11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/status": {
            "put": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "[Admin] Cập nhật trạng thái đơn hàng",
                "operationId": "2c671548bfccd7f6ddbc7ea21773ed61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "processing",
                                            "shipped",
                                            "delivered",
                                            "cancelled",
                                            "returned"
                                        ]
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "failed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT token vào đây để thực hiện các yêu cầu có bảo mật.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}